+++ packages/rabbitmq/src/amqp/connection.ts
@@ -184,9 +184,15 @@ export class AmqpConnection {
           return;
         }
 
+        // Check that the Buffer has content, before trying to parse it
+        const message =
+          msg.content.length > 0
+            ? JSON.parse(msg.content.toString())
+            : undefined;
+
         const correlationMessage: CorrelationMessage = {
           correlationId: msg.properties.correlationId.toString(),
-          message: JSON.parse(msg.content.toString()),
+          message: message,
         };
 
         this.messageSubject.next(correlationMessage);
@@ -210,10 +216,15 @@ export class AmqpConnection {
       first()
     );
 
-    this.publish(requestOptions.exchange, requestOptions.routingKey, payload, {
-      replyTo: DIRECT_REPLY_QUEUE,
-      correlationId,
-    });
+    await this.publish(
+      requestOptions.exchange,
+      requestOptions.routingKey,
+      payload,
+      {
+        replyTo: DIRECT_REPLY_QUEUE,
+        correlationId,
+      }
+    );
 
     const timeout$ = interval(timeout).pipe(
       first(),
@@ -247,18 +258,7 @@ export class AmqpConnection {
     msgOptions: MessageHandlerOptions,
     channel: amqplib.ConfirmChannel
   ): Promise<void> {
-    const { exchange, routingKey, allowNonJsonMessages } = msgOptions;
-
-    const { queue } = await channel.assertQueue(
-      msgOptions.queue || '',
-      msgOptions.queueOptions || undefined
-    );
-
-    const routingKeys = Array.isArray(routingKey) ? routingKey : [routingKey];
-
-    await Promise.all(
-      routingKeys.map((x) => channel.bindQueue(queue, exchange, x))
-    );
+    const queue = await this.setupQueue(msgOptions, channel);
 
     await channel.consume(queue, async (msg) => {
       try {
@@ -271,7 +271,7 @@ export class AmqpConnection {
         const response = await this.handleMessage(
           handler,
           msg,
-          allowNonJsonMessages
+          msgOptions.allowNonJsonMessages
         );
 
         if (response instanceof Nack) {
@@ -332,18 +332,7 @@ export class AmqpConnection {
     rpcOptions: MessageHandlerOptions,
     channel: amqplib.ConfirmChannel
   ) {
-    const { exchange, routingKey, allowNonJsonMessages } = rpcOptions;
-
-    const { queue } = await channel.assertQueue(
-      rpcOptions.queue || '',
-      rpcOptions.queueOptions || undefined
-    );
-
-    const routingKeys = Array.isArray(routingKey) ? routingKey : [routingKey];
-
-    await Promise.all(
-      routingKeys.map((x) => channel.bindQueue(queue, exchange, x))
-    );
+    const queue = await this.setupQueue(rpcOptions, channel);
 
     await channel.consume(queue, async (msg) => {
       try {
@@ -356,7 +345,7 @@ export class AmqpConnection {
         const response = await this.handleMessage(
           handler,
           msg,
-          allowNonJsonMessages
+          rpcOptions.allowNonJsonMessages
         );
 
         if (response instanceof Nack) {
@@ -372,7 +361,7 @@ export class AmqpConnection {
           this.logger.verbose(
             `reply to is enabled, sending response with correlationId ${correlationId}`
           );
-          this.publish('', replyTo, response, { correlationId });
+          await this.publish('', replyTo, response, { correlationId });
         }
         channel.ack(msg);
       } catch (e) {
@@ -443,5 +432,55 @@ export class AmqpConnection {
     }
 
     return handler(message, msg);
+  }
+
+  private async setupQueue(
+    subscriptionOptions: MessageHandlerOptions,
+    channel: amqplib.ConfirmChannel
+  ): Promise<string> {
+    const {
+      exchange,
+      routingKey,
+      createQueueIfNotExists = true,
+    } = subscriptionOptions;
+
+    let actualQueue: string;
+
+    if (createQueueIfNotExists) {
+      const { queue } = await channel.assertQueue(
+        subscriptionOptions.queue || '',
+        subscriptionOptions.queueOptions || undefined
+      );
+      actualQueue = queue;
+    } else {
+      const { queue } = await channel.checkQueue(
+        subscriptionOptions.queue || ''
+      );
+      actualQueue = queue;
+    }
+
+    let bindQueueArguments: any;
+    if (subscriptionOptions.queueOptions) {
+      bindQueueArguments = subscriptionOptions.queueOptions.bindQueueArguments;
+    }
+
+    const routingKeys = Array.isArray(routingKey) ? routingKey : [routingKey];
+
+    if (exchange && routingKeys) {
+      await Promise.all(
+        routingKeys.map((routingKey) => {
+          if (routingKey != null) {
+            channel.bindQueue(
+              actualQueue as string,
+              exchange,
+              routingKey,
+              bindQueueArguments
+            );
+          }
+        })
+      );
+    }
+
+    return actualQueue;
   }
 }